Fixed a number of problems:
The key issue was that you were instantiating a view controller from within a view controller. You generally don't do that (and if you do want to do that, you do it differently than you did). When I fixed that (getting rid of ProfilTableView in the process), your problem went away.

A few unrelated changes:

- Variables always start with lowercase letters; types (such as your `CellData` struct`) always start with uppercase letter. So I changed a bunch of variables and types accordingly.
- I eliminated your global variables. Only use globals if you absolutely need to. You don't here.
- You were adding your refresherControl twice. I got rid of one of those.